comfy/custom_nodes/ComfyUI_LayerStyle_Advance/requirements.txt
numpy
matplotlib
scikit_image
scikit_learn
opencv-contrib-python
pymatting
timm
blend_modes
transformers>=4.45.0
diffusers
loguru
colour-science
huggingface_hub>=0.23.4
segment_anything
addict
omegaconf
yapf
wget
iopath
mediapipe
typer_config
fastapi
rich
google-generativeai
google-genai>=1.5.0
pillow>=10.1.0
ultralytics>=8.2.0
transparent-background
accelerate>=0.26.0
onnxruntime
bitsandbytes>=0.41.1
peft>=0.12.0
protobuf
hydra-core
blind-watermark
qrcode
pyzbar
psd-tools
zhipuai
openai
typer-config
fastapi